Move splash screen to separate files
