+++ app/src/main/java/com/realityexpander/tasky/presentation/splash_screen/SplashState.kt
@@ -1,0 +1,10 @@
+package com.realityexpander.tasky.presentation.splash_screen
+
+import com.realityexpander.tasky.domain.AuthInfo
+import com.realityexpander.tasky.presentation.util.UiText
+
+data class SplashState(
+    val authInfo: AuthInfo? = null,
+    val authInfoChecked: Boolean = false,
+    val statusMessage: UiText = UiText.None,
+)
